Common Datamodule. Fixes #5
